requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
boto3==1.22.5
boto3-type-annotations==0.3.1
botocore==1.25.5
certifi==2021.10.8
colorama==0.4.4
dacite==1.6.0
elastic-apm==6.9.1
etcd3==0.12.0
greenlet==1.1.2
grpcio==1.46.0
iniconfig==1.1.1
jmespath==1.0.0
lxml==4.8.0
packaging==21.3
pika==1.2.1
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.36
tenacity==8.0.1
tomli==2.0.1
urllib3==1.26.9
